[refac] Refactor sign-up process to use new token service and response structure
