+++ edukit-api/build.gradle
@@ -20,11 +20,6 @@ dependencies {
 	// Database
 	implementation 'com.mysql:mysql-connector-j'
 
-	// JWT
-	implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-	implementation 'io.jsonwebtoken:jjwt-impl:0.11.5'
-	implementation 'io.jsonwebtoken:jjwt-jackson:0.11.5'
-
 	// Health Check
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 }
