+++ edukit-api/src/main/java/com/edukit/api/controller/auth/AuthController.java
@@ -2,14 +2,11 @@ package com.edukit.api.controller.auth;
 
 import com.edukit.api.common.ApiResponse;
 import com.edukit.api.controller.auth.request.MemberSignUpRequest;
-import com.edukit.api.controller.auth.response.MemberSignUpResponse;
 import com.edukit.api.security.handler.RefreshTokenCookieHandler;
-import com.edukit.api.security.jwt.service.JwtGenerator;
-import com.edukit.api.security.jwt.service.Token;
 import com.edukit.api.security.util.PasswordValidator;
 import com.edukit.common.exception.code.CommonSuccessCode;
 import com.edukit.core.auth.facade.AuthFacade;
-import com.edukit.core.auth.facade.dto.SignUpResult;
+import com.edukit.core.auth.facade.response.MemberSignUpResponse;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -25,7 +22,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthController {
 
     private final AuthFacade authFacade;
-    private final JwtGenerator jwtGenerator;
     private final PasswordEncoder passwordEncoder;
     private final RefreshTokenCookieHandler cookieHandler;
 
@@ -34,12 +30,10 @@ public class AuthController {
                                                     final HttpServletResponse servletResponse) {
         PasswordValidator.validatePasswordFormat(request.password());
         String encodedPassword = passwordEncoder.encode(request.password());
-        SignUpResult result = authFacade.signUp(request.email(), encodedPassword, request.subject(), request.nickname(),
-                request.school());
+        MemberSignUpResponse response = authFacade.signUp(request.email(), encodedPassword, request.subject(),
+                request.nickname(), request.school());
 
-        Token tokens = jwtGenerator.generateTokens(result.memberUuid());
-        MemberSignUpResponse response = MemberSignUpResponse.of(tokens.accessToken());
-        servletResponse.addCookie(cookieHandler.createRefreshTokenCookie(tokens.refreshToken()));
+        servletResponse.addCookie(cookieHandler.createRefreshTokenCookie(response.refreshToken()));
         return ApiResponse.success(CommonSuccessCode.OK, response);
     }
 }
