+++ edukit-core/build.gradle
@@ -10,6 +10,11 @@ dependencies {
 
 	// Jackson
 	implementation 'com.fasterxml.jackson.core:jackson-databind'
+
+	// JWT
+	implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+	implementation 'io.jsonwebtoken:jjwt-impl:0.11.5'
+	implementation 'io.jsonwebtoken:jjwt-jackson:0.11.5'
 }
 
 // 라이브러리 JAR 생성
