+++ edukit-core/src/main/java/com/edukit/core/auth/facade/AuthFacade.java
@@ -2,9 +2,11 @@ package com.edukit.core.auth.facade;
 
 import com.edukit.core.auth.enums.AuthCodeType;
 import com.edukit.core.auth.event.MemberSignedUpEvent;
-import com.edukit.core.auth.facade.dto.SignUpResult;
+import com.edukit.core.auth.facade.response.MemberSignUpResponse;
 import com.edukit.core.auth.service.AuthCodeService;
 import com.edukit.core.auth.service.AuthService;
+import com.edukit.core.auth.service.Token;
+import com.edukit.core.auth.service.TokenService;
 import com.edukit.core.member.entity.Member;
 import com.edukit.core.member.service.MemberService;
 import com.edukit.core.subject.entity.Subject;
@@ -22,17 +24,22 @@ public class AuthFacade {
     private final AuthCodeService authCodeService;
     private final MemberService memberService;
     private final SubjectService subjectService;
+    private final TokenService tokenService;
     private final ApplicationEventPublisher eventPublisher;
 
     @Transactional
-    public SignUpResult signUp(final String email, final String password, final String subjectName,
-                               final String nickname, final String school) {
+    public MemberSignUpResponse signUp(final String email, final String password, final String subjectName,
+                                       final String nickname, final String school) {
         authService.validateCondition(email, nickname);
         Subject subject = subjectService.getSubjectByName(subjectName);
         Member member = memberService.createMember(email, password, subject, nickname, school);
+
+        Token token = tokenService.generateTokens(member.getMemberUuid());
+        // refreshToken을 Redis에 저장하는 로직 구현
+
         String authCode = authCodeService.issueVerificationCode(member, AuthCodeType.TEACHER_VERIFICATION);
 
         eventPublisher.publishEvent(MemberSignedUpEvent.of(member.getEmail(), member.getMemberUuid(), authCode));
-        return SignUpResult.of(member.getMemberUuid());
+        return MemberSignUpResponse.of(token.accessToken(), token.refreshToken());
     }
 }
