--- edukit-core/src/main/java/com/edukit/core/auth/facade/dto/SignUpResult.java
@@ -1,9 +1,0 @@
-package com.edukit.core.auth.facade.dto;
-
-public record SignUpResult(
-        String memberUuid
-) {
-    public static SignUpResult of(final String memberUuid) {
-        return new SignUpResult(memberUuid);
-    }
-}
