+++ edukit-core/src/main/java/com/edukit/core/auth/jwt/provider/JwtKeyProvider.java
@@ -1,4 +1,4 @@
-package com.edukit.api.security.jwt.provider;
+package com.edukit.core.auth.jwt.provider;
 
 import io.jsonwebtoken.security.Keys;
 import java.security.Key;
