+++ edukit-core/src/main/java/com/edukit/core/auth/jwt/setting/JwtPropertiesConfig.java
@@ -1,4 +1,4 @@
-package com.edukit.api.security.jwt.setting;
+package com.edukit.core.auth.jwt.setting;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
